Fixed a major casting bug with ExtendedBinaryReader. Basically the bitwise left/right shift operators in C# can only take the following types on it's left-hand side: - int - uint - long - ulong
If you pass in any other type of data, it'll just auto-cast it to one of those types (usually int) if able, otherwise you won't be able to compile.

WELL basically we were passing in a byte... which was getting auto-casted to an int.
This means that a function like ReadDouble was essentially casting each byte to an int first, THEN bitshifting the int value, then casting that to a ulong, rather than just casting to a ulong and bitshifting that.

This meant certain larger values were getting read entirely incorrectly. Really sorry!