Huge improvements to ManiaSprite.cs Found out a lot of the unknown stuff! I was right - the stuff after each frame entry actually WAS related to the positioning of the sprite.
Turns out it's actually the origin X and Y! Somehow I didn't get that the value was signed lol so I just had no idea what it was.

Also found a sprite that actually has multiple textures (Data/Sprites/Eggman/EggmanAll.bin). Turns out I got the structure a bit wrong! Also I found that the first byte in the unknown part at the beginning of each frame is an index to the texture used by that frame.