HedgeArcPack alpha 0.1 Can extract and repack LW archives, big and little endian, with default split size or a custom one (which can be specified in decimal, hex, or octal). Flags can also be auto-determined based on input type, meaning you can simply drag-n-drop LW pacs (even splits) onto the program to extract them.
Archive type cannot yet be auto-determined from a directory when packing, however, so the user is prompted to pick what type of archive they'd like to create when dragging a folder onto the program. In future versions I plan on improving this by having HedgeArcPack determine the archive type from archives with the same name as the folder (e.g. "Sonic.pac" when the user is trying to pack the "Sonic" folder), meaning you'd only have to specify the archive type when packing a folder for the first time; upon repacking the type would be determined automatically. I could also guess the input type from contents of the folder, etc. though I doubt I'll go that route as I don't wish to over-complicate things. If it isn't obvious to the program what type of archive you wish to pack, it should just ask imo.