Lots of new functions for dealing with Archives Most importantly there's hl_ExtractArchivesOfType which lets you manually specify what type the archives are instead of having it auto-detected if you wish, and hl_GetArchiveSplitCount/hl_GetArchiveSplitCountRoot which replace the old hl_ArchiveGetSplits function.
I realized after making that function that it was really dumb to have it return an array of strings for a multitude of reasons, the biggest of which being how much unnecessary memory/CPU that wastes. It makes much more sense to just return the split count imo. This new function also only requires a filePath, not a loaded archive.