Unicode <-> CP-932 (aka "Shift-JIS") conversions writing this killed me and broke my spine No but seriously this was far more of a pain than it needed to be.
BINA files appear to treat strings as merely a sequence of bytes, meaning they actually accept any encoding as long as 00 is only ever used as the null-terminator.
Because of this, some (or maybe all?) BINA files actually contain text encoded in Windows codepage 932 (aka "CP-932" or "Shift-JIS", although this is somewhat inaccurate as "Shift-JIS" is actually a separate encoding CP-932 was based off of; they're mostly the same but contain some major differences).

An example of this is the Forces .gedit files which contain ObjDebugNotes (basically "comments").
These objects reference strings encoded in CP-932.

On Windows I can simply convert from the 932 codepage to Unicode, but on other platforms it's not so simple as these all have their own codepages/locales which have some differences.
0x7E in standard Shift-JIS, for example, is an overline (U+203E ‾), whereas in CP-932 it's a tilde (U+007E ~).
I wanted something that worked exactly the same cross-platform, so I wrote this.

This code works using mapping tables in INCP932Mappings.h, which was generated by a quick and simple tool I wrote using Shawn Steele's "cp932 to Unicode table" (available here: http://unicode.org/Public/MAPPINGS/VENDORS/MICSFT/WINDOWS/CP932.TXT).

Given I'm not an expert when it comes to encodings, this code is *very messy* and could probably be more efficient - but it works. Or at least it did in my testing.
I wrote a quick text file containing various Kanji/Hiragana/Katakana/ASCII characters in CP-932 and these functions were able to properly convert it to both UTF-8 and UTF-16.
I was also able to convert the resulting UTF-8/UTF-16 back to CP-932 and got the expected result. This worked on both Windows and Linux.