please don't yell at me, I already know. Screw C++ lol, I'm going back to C. Please hear me out.
I've never been happy with my C++ code somehow. I always thought it was my fault. I never had this problem back when I used C# as my primary language, but I assumed that was just due to me changing a lot since then and having higher standards now.

Turns out it was just C++.

I've been using C almost exclusively for several months now and I've felt something I haven't felt while programming for a while now: happiness lol.

In C I'm in control of everything; there's no hidden nightmare behaviour that happens behind the scenes like is constantly the case in C++. Plus a lot of crap is just easier.
Originally I ended up typing an entire rant here trying to explain what exactly I meant by that which was literally like 6 or 7 paragraphs long LOL and uhh, on second thought, nah I don't think anyone wants to read all of that.

I guess to summarize: I could rant about C++ for literal hours if I wanted to. I don't think it's a bad language, actually. In fact, despite everything I just said, I think it can be great if used correctly. But I've found that getting to that point where C++ becomes great requires a lot of dealing with the language's crap, and this generally tends to be a very time-consuming and stressful task in and of itself. I've found that I'm actually much happier not having to do any of that at all and being able to just write code and have it work well.

Believe it or not, C lets me do just that.

C has data. It has functions. That's about it lol. And that's great!! That's all I need.
C doesn't make me constantly worry about a bunch of implicit complicated rules; I just write code and it does what it looks like it does.

Not to mention, it's ridiculously portable - even more so than C++ (I wasn't able to compile HedgeLib on macOS for a while, for example, due to some weird obscure difference with how the macOS version of clang handles template specialization that I'm not sure is even standard C++) - and it compiles much faster due to just being a much simpler language.

The only "good practices" anyone expects in C are as follows:
- Make it work.
- Make it safe.
- Make it clean.
- Make it fast.

If you do those no one really cares about the rest. This is in stark contrast to C++ where every little detail of everything you do has multiple ways it could be done, one of which is always considered to be better (even if it isn't), and if you don't do it exactly that way all the time, unless you have a very good reason (read: you literally have to do it differently), you're writing code that follows "bad practices". "Why are you using new?!! You have make_unique!! What's wrong with you??!??" "BRUH why'd you use a char* here?! USE std::string !! ARE YOU CRAZY?!" Stuff like this.

Basically out of all the projects I was working on, I found that it was only HedgeLib that was bringing me distress still. I didn't want to touch it because every time I did I felt like I was wasting my time trying to fix the unfixable mess of C++. I obviously didn't want to fricking rewrite stuff again, but given all of this, the only way it was going to really get better was if I did. I experimented with a pure C89 version of HedgeLib for a month or two. And dang, unlike with all the previous "rewrites", I actually felt happy again while writing it. I don't feel like I'm writing "meh code" all the time while working on this; I feel like I'm actually writing something solid that can last. This commit is what I have so far of that C89 version.

I'm still going to keep the branch name as "HedgeLib++" for now just because I don't know what else to call it lol and intend to finally get my act together and merge this into the master branch soon anyway.

I also haven't completely lost my mind: I'm still going to use C++ where it makes sense.

HedgeEdit, for example, will still be written in C++ because it's using Qt which I believe only supports C++. It's also doing Direct3D stuff which can actually be done from C... but it's a nightmare lol and not at all worth it. So yeah I'll be sticking with C++ for HedgeEdit. It's been removed temporarily from the repo though cuz it currently doesn't compile.

I'll be keeping all of the HedgeEdit code that I feel was good lol, like how I was handling various formats/games. The rendering backend, though, really did need some serious reworking. I wasn't even doing any sorting or anything and the architecture I was using didn't really allow for it. I've been working on a rendering engine as a side-project for a month or so now to teach me various things. I plan on using this engine in HedgeEdit once it's ready, which will be soon. Most of the architecture and such is already done, and it's *much much* better than what I had. It's also been designed in a forward-thinking way so I shouldn't ever have to rewrite it lol. For example: it's actually entirely designed around modern APIs such as d3d12 and vulkan, since I firmly believe this is where graphics APIs will be heading in the future. In spite of this, it translates shockingly well to older APIs such as d3d11 and GL, so once this is done, I'll be able to can simply write multiple backends and use whichever backend I feel is most appropriate for the target platform, all with relative ease.

I'm also finally going to be committing much more frequently from here on out. I need to yeet this mindset I have that committing to github == releasing a stable build. Obviously that's not fricking true and I know that, but it's hard to get it out of my head. Imma attempt to force myself to git push from now on so hopefully that'll work out lol.

To anyone who read all of this: Thanks a ton for dealing with my crap. I'm sorry for constantly doing this; I suppose you could say I've simply been in pursuit of happiness for a while. I feel as though I've finally actually found it, though. I plan to never rewrite this fricking library again from here on out and *actually finish stuff*. Hopefully I'll actually be successful this time.