Fixes for 32-bit Windows
It appears Windows character conversion functions internally verify not only that the converted string will fit within the provided destination buffer size, but also that you **HAVE ENOUGH MEMORY FREE** to safely allocate a **NEW** buffer of that size even though **THE FUNCTION DOESN'T FRICKING ALLOCATE ANYTHING**?!!?

This is an undocumented quirk (honestly it seems like a bug) in the Win32 API.

In our HedgeLib conversion functions, we allow the user to not specify the destination buffer size (by just passing in 0) if they don't care. Since these functions call the Windows ones internally which require a non-zero destination buffer size, I just used INT_MAX as the destination buffer size if it wasn't given by the user.

Well, on 32-bit Windows, because of the aforementioned quirk, this will always result in ERROR_INSUFFICIENT_BUFFER, as 32-bit Windows does not allow you to allocate more than 2GB of memory through normal means (2GB == 2000000000 and INT_MAX == 2147483647).

We could just use a smaller "large number" but then we're just playing a guessing game everytime we call a Win32 character conversion function ("Will it work this time or nah?") as whether or not it will work is entirely reliant upon how much memory the user currently has allocated.

Best we can do without changing our API to require the user to specify the destination buffer size is to simply compute the required length for the conversion and use that as the destination buffer size if the user doesn't provide one themselves. This is undesirable as we've often already computed the length beforehand, so we're doing (potentially slow) work twice for no reason. But it should work everytime.

Thanks Microsoft. :)