rad push HedgeLib ok skyth <3
So if you somehow didn't know, yes, I rewrote it again. For the last time this time. Seriously. I promise.

The decision to switch to C was made in a bad time in my life when I was in a really bad mental state and honestly it just didn't make any sense.
I just wanted something fresh and new to work on to help better distract me from stuff going on in my life at the time. That was the only reason I did it.

Problem was, it wasn't just a silly decision without a good reason behind it, it was an actively bad decision that actually added a ton of completely unnecessary coding work and slowed development down by a HUGE amount completely unnecessarily.
So I've rewritten it one final time in C++.

From here on out I intend to:
- ACTIVELY commit and push (aka "stop treating pushes as releases")
- NEVER rewrite HedgeLib again

I also plan to adopt a new style suggested by Skyth that's actually somewhat similar to C# HedgeLib while still allowing direct loading and such if you want to do that.
This style is already in-use in the new HH model/material/texture stuff and the HL scene stuff present in this commit.

The shader stuff in this commit is not yet in this style and will be changed to match it soon. I wasn't going to commit this all yet due to some things being unfinished like this, but skyth, death, and others highly encouraged me to do so as they rightfully pointed out I've been treating my commits as releases and need to stop doing that. It compiles. That's good enough for a commit (many projects don't even care about that!).