Fix compilation on macOS
For some reason it seems the compiler used in this version of macOS doesn't defined aligned_alloc in the std namespace. It does seem to define it in the global namespace, though?

The more standards-compliant solution I think is to just fallback to the POSIX approach on platforms that have __APPLE__ defined for now.