Add aligned read/write to streams/writers
You can now say, for example:
stream.write_obj(myObj, 16);

This will write myObj to the stream with a byte alignment of 16.
It effectively pads the stream to the given alignment, then writes the data.

You can also read data with a given alignment, like so: stream.read_obj(myObj, 16);

This also works with read_arr and write_arr.

If you specify an alignment of 0 to any of these new functions, they will instead use an alignment of alignof(T). So, for example:

s32 myVal = 7;
stream.write_obj(myVal, 0);

This will pad the stream to an alignment of 4 (since alignof(s32) is 4), then write the value to the stream.

To read/write without alignment, simply don't specify an alignment argument when calling any of these functions, or specify an alignment of 1.