PACx: Fix incorrect offsets being "fixed" on write
I'm working on a rewrite for the PACx stuff, but in the meantime, this should fix some of the corrupted pac file generation bugs people have been reporting. It's not going to fix everything, as I'm currently aware of at least one other bug causing corrupted pac files to be generated that I still have yet to fix.