HedgeSet: Attempt to fix random rotation bug
Bizarre bug. A random (jibberish) rotation value is sometimes written to generated gedits for unknown reasons. It seems to be triggered by attempts to get the global rotation, but *not* by attempts to get the local rotation. My local builds seemed fine, so this issue also seems exlusive to the AppVeyor builds?! Perhaps due to using a different version of DirectXMath, or due to using a different version of MSVC? I have no idea, but I'm guessing it might be caused by the quaternion alignment trick I do in this function? This commit attempts to remove that hack to see if it fixes the bug.