Handle non-SRT matrix4x4 decompose
Wow, if DirectXMath thinks this is a non-SRT matrix, it outputs a valid position and scale, but doesn't touch the rotation value at all, and returns false, meaning in my case, if DirectXMath determined it was dealing with a non-SRT matrix, it would leave me with uninitialized rotation data!

To fix this, I now check the return value, and return it back tot he user as well, who can decide to do what they want to with it.