BINA: Simplify fix functions.
Previously, to use data from a BINA file, you first had to call fix32/fix64 to fix the BINA container data, then get_data<T> to get a pointer of type T to the data within the BINA container, then finally, data->fix(...) to fix the data contained within the BINA container.

Now, you can do all of that with a single call to fix32/fix64.

I want to also make it so endianFlag is automatically passed in as the first parameter to the fix function, but for that to work cleanly I also have to take the pacpack metadata stuff out of the BINA fix functions.