BINA: Use relative offset positions in iterators
Previously we returned the position of each offset, relative to the base. This just wasn't a very good abstraction for this use-case. Now, we just return the position of each offset, relative to the position of the previous offset. This is how the data is actually stored in the BINA offset table anyway, so it makes for a simpler implementation, and it's trivial to modify existing code to turn that into the absolute offset position like we were returning before.