Add Forces HSON Template
This is a WIP; mainly due to the lack of descriptions. This is because the descriptions present in the PC version of the game unforuntately are corrupted, presumably due to the developers compiling the Shift-JIS-encoded source code on a machine with English locale, thus causing MSVC to treat the file's encoding as ANSI, rather than Shift-JIS. The result is that all of the Japanese characters are replaced with ?? characters. The console versions do not have this corruption, thus it's possible to extract the Japanese descriptions from them. However, doing so would be a massive pain, and the descriptions are usually nothing more than a direct translation of the parameter names to Japanese anyway (seriously not useful). Instead, what I've done is I simply extracted every description from the PC version that does not contain any corrupted ?? characters (the ones with all-English-characters). As you can see, they're not very useful. In the future, I would like to replace these with custom descriptions, such as the very helpful ones written by Duck Dealer for C# HedgeEdit.